Frontend search view

+++ src/main/kotlin/com/example/myHomeAppliance2/Id.kt
@@ -1,3 +1,3 @@
 package com.example.myHomeAppliance2
 
-data class Id(val id : Long)
+data class Id(val id : Int)
